Defined file structure for the ETL processes.
I created a series of empty Python scripts in order to define the
file structure of the various data ingestion scripts. The files can
be found under `src/` and have been organized like so:

Topic #1
	"Data Source #1".py
	"Data Source #2".py
Topic #2
	"Data Source #1".py
Topic #3
	"Data Source #1".py
[...]

This file structure merely adheres to the current expectations of
the project and may be subject to change. I find the current
structure to be the easiest to traverse as it mirrors the vision of
the project.