Added `data` directory.
This is where all the generated data will be stored. My one gripe is
that one may need to link to a deep file traversal to get from `src`
to `data`.

For example, say one saves a spreadsheet from
"src/topics/hourly_fuel_output/ieso.py", one needs to link to
"../../data" in order to save their file.

Regardless, I find this to be the most straight forward and likely
to change. Besides, I'm not sure where to save the data. It may be
on the cloud. This is merely a temporary, local solution.