Added more libraries to install.
Google Collab installs a LOT of libraries for you on default. As a
result, when it moved locally, I had to reinstall a lot of
libraries. Note that I am using a virtual enviorment in a completely
different directory. If you want to reproduce this code, I would
recommend that you create one.